requirements/base.txt
Flask
Flask-SQLAlchemy
sqlalchemy-migrate
